Modified builds
Modified builds.

+++ messageStep2/clean.sh
@@ -1,3 +1,5 @@
 #!/bin/sh
 
-rm -r mods
+if [ -d "mods" ]; then
+	rm -r mods
+fi
